+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.6] - 2026-06-27
+
+### Added
+- `activeUrl` and `activeId` parameters to the `AppPage` to handle the late addtion of navigation items
+
+### Changed
+- `AppPage` and `useTreeNavigation` variable names to use focused instead of active
+
 ## [0.12.5] - 2026-06-26
 
 ### Added